Add files via upload
This is a simple .bat script written in PowerShell that simplifies running the system file checker, deleting all temporary files in the Temp directory, and upgrading needed OS software using winget.

I initially put three commands into a .bat file, as pictured. However, I wanted to try and use some material I learned from my PowerShell scripting class, adding a UI, along with a permission check to make sure the script is run by an Administrator for the System File Checker to run correctly. I'd like to make further improvements where I can on this script, but this is a wonderful start on my journey creating scripts for basic purposes, and eventually for cybersecurity and automation.